update routing pages, add metadata
