+++ src/types.ts
@@ -1,4 +1,4 @@
-interface Task {
+export interface Task {
   id: number;
   title: string;
   description: string;
